Moved date replacement to RunnableBase
